// types/api/Pokemons.ts
export interface GenderRatio {
  male: number
  female: number
  isGenderless: boolean
};